+++ src/Visos/Graphics/PvZBuffSurface.cpp
@@ -39,7 +39,7 @@ void PvZBuffSurface::AllocateZBuff()
 
 	size.m_width = (short) (m_windowRect.m_width * 2);
 	size.m_height = m_windowRect.m_height;
-	size = m_bitmap.SetSize(size, (int) m_reserved40 * 2);
+	size = m_bitmap.SetSize(size, m_reserved40 * 2);
 	allocatedArea = (unsigned int) m_allocatedWidth * (unsigned int) m_allocatedHeight * 2;
 	neededArea = (unsigned int) ((int) size.m_height * (int) size.m_width);
 	if (allocatedArea < neededArea) {
